pyproject.toml
[project]
name = "agents"
version = "0.0.1"
description = "Agents for LangGraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "browserbase>=1.4.0",
    "langgraph>=1.0.0",
    "langgraph-sdk>=0.2.10",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["pyrefly>=0.44.1", "ruff>=0.6.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "pyrefly>=0.44.1",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
